Modified item-macros.cmake to remove the XXD fallback. It is incompatible now.
Stylistic changes:
   base64.hpp moved to KIM_Base64.hpp
   KIM_Base64.hpp edited to have 2 space indent
   Function names/variable names to follow PascalCase or camelCase (like other KIM-API files)
   Doxygen style docstrings
TODO:
   Licensing issue discussion with Ryan
