+++ cmake/items-macros.cmake.in
@@ -530,30 +530,34 @@ function(_add_custom_command_blob_to_cpp _dirin _filein _fileout)
     unset(_cfilein)
     unset(_edit_xxd_output)
   else()
-    find_package_message(xxd "Missing xxd: Falling back to less efficient cmake implementation." "missing")
+    # No backup yet
+    # Possible Solutions: 1) CMAKE only implementation of base64 encoding? 
+    #                     2) use system base64/openssl binaries?
+    message(FATAL_ERROR "Missing kim-base64-encode: Please check the KIM-API installation.")
+    # find_package_message(xxd "Missing xxd: Falling back to less efficient cmake implementation." "missing")
 
-    string(MAKE_C_IDENTIFIER ${_filein} _cfilein)
-    set(_blob_to_source "${CMAKE_CURRENT_BINARY_DIR}/BlobToCSource.cmake")
-    if(NOT EXISTS "${_blob_to_source}")
-      file(WRITE "${_blob_to_source}"  # use a bracket argument to avoid ugly escaping
-        [=[
-        # This file was automatically generated by CMake; do not edit.
-        file(READ "${filein}" _content HEX)
-        string(REGEX REPLACE "([0-9a-f][0-9a-f])" "0x\\1," _content "${_content}")
-        string(REGEX REPLACE ",$" "" _content "${_content}")
-        file(WRITE  "${fileout}" "unsigned char const ${varname}[] = { ${_content} };\n")
-        file(APPEND "${fileout}" "unsigned int const ${varname}_len = sizeof( ${varname} );\n")
-        unset(_content)
-        ]=]
-        )
-    endif()
-    add_custom_command(OUTPUT ${_fileout}
-      COMMAND ${CMAKE_COMMAND} "-Dvarname=${_cfilein}" "-Dfilein=${_filein}" "-Dfileout=${_fileout}" -P "${_blob_to_source}"
-      DEPENDS "${_dirin}/${_filein}" "${_blob_to_source}"
-      WORKING_DIRECTORY "${_dirin}"
-      )
-    unset(_cfilein)
-    unset(_blob_to_source)
+    # string(MAKE_C_IDENTIFIER ${_filein} _cfilein)
+    # set(_blob_to_source "${CMAKE_CURRENT_BINARY_DIR}/BlobToCSource.cmake")
+    # if(NOT EXISTS "${_blob_to_source}")
+    #   file(WRITE "${_blob_to_source}"  # use a bracket argument to avoid ugly escaping
+    #     [=[
+    #     # This file was automatically generated by CMake; do not edit.
+    #     file(READ "${filein}" _content HEX)
+    #     string(REGEX REPLACE "([0-9a-f][0-9a-f])" "0x\\1," _content "${_content}")
+    #     string(REGEX REPLACE ",$" "" _content "${_content}")
+    #     file(WRITE  "${fileout}" "unsigned char const ${varname}[] = { ${_content} };\n")
+    #     file(APPEND "${fileout}" "unsigned int const ${varname}_len = sizeof( ${varname} );\n")
+    #     unset(_content)
+    #     ]=]
+    #     )
+    # endif()
+    # add_custom_command(OUTPUT ${_fileout}
+    #   COMMAND ${CMAKE_COMMAND} "-Dvarname=${_cfilein}" "-Dfilein=${_filein}" "-Dfileout=${_fileout}" -P "${_blob_to_source}"
+    #   DEPENDS "${_dirin}/${_filein}" "${_blob_to_source}"
+    #   WORKING_DIRECTORY "${_dirin}"
+    #   )
+    # unset(_cfilein)
+    # unset(_blob_to_source)
   endif()
 endfunction(_add_custom_command_blob_to_cpp)
